src/phoenix_starter_kits/templates/langchain-rag/{{cookiecutter.project_slug}}/requirements.txt
arize-phoenix[evals]>=3.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
faiss-cpu>=1.7.4
openinference-instrumentation-langchain>=0.1.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0